% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{acf_trawl}
\alias{acf_trawl}
\title{ACF of Trawl processes}
\usage{
acf_trawl(object, h, dff = 0, lag_max = 25, drop_zero = TRUE)
}
\arguments{
\item{\dots}{any other passthru pareters}
}
\description{
theoretical autocorrelation of a trawl process
}
\details{
TODO

CITE TODO.
}
\examples{

TODO

# simulations estimation
TODO

}
\references{
TODO
}
\seealso{
\code{\link{fit_trawl}}
}
\author{
Dries Cornilly
}
\concept{
trawl
}
