% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESM8}
\alias{ESM8}
\title{Trade prices in ES}
\format{An xts object with 1 variable
\describe{
  \item{Price}{trade price at which a trade occurred, in USD}
}}
\source{
Brian G. Peterson
}
\usage{
ESM8
}
\description{
A dataset containing the trade prices for the ES contract with expiration
M8 on May 7, 2018 between 8h30 and 12h00 CDT.
}
\keyword{datasets}
