% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trawl.R
\name{sim_trawl}
\alias{sim_trawl}
\title{Simulation of Trawl processes}
\usage{
sim_trawl(object, univariate = TRUE, ...)
}
\arguments{
\item{object}{bla}

\item{univariate}{bla}

\item{\dots}{any other passthru pareters}
}
\description{
simulates a path of a Trawl process
}
\details{
TODO

CITE TODO.
}
\examples{

#TODO

# simulations estimation
#TODO

}
\references{
TODO
}
\seealso{
\code{\link{fit_trawl}}
}
\author{
Dries Cornilly
}
\concept{
trawl
}
