% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_trawl.R
\encoding{UTF-8}
\name{acf_trawl}
\alias{acf_trawl}
\title{Autocorrelation function}
\usage{
acf_trawl(object, h, method = "sample", dff = 0, lag_max = 25,
  drop_zero = TRUE, ...)
}
\arguments{
\item{object}{object containing all the specifications for the process, see details}

\item{h}{observation frequency}

\item{method}{"sample" for the empirical values, something else for the theoretical values}

\item{dff}{number of times the series has to be differenced before computing the 
autocorrelation, possible values are 0 (default) and 1}

\item{lag_max}{number of lags to consider, default 25}

\item{drop_zero}{whether to exclude the autocorrelation at lag 0, default TRUE}

\item{\dots}{any other passthrough parameters}
}
\description{
compute empirical or theoretical autocorrelation values of either the process
itself or the process of first differences
}
\details{
the object is either an output object from sim_trawl for the empirical autocorrelation
values, or an output object from fit_trawl for the theoretical ones. When computing the sample
ACF, it is possible to add the argument 'multi' for reducing the 
estimation variance.

The sample method works for any type of time series, regularly or irregularly spaced.
}
\examples{
# empirical and estimated autocorrelations
sim <- sim_trawl(list())
sim$h <- 0.5
sim$trawl <- "exp"
sim$lag_max <- 3
sim$method <- "acf"
ft <- fit_trawl(sim)
plot(1:10, acf_trawl(sim, sim$h, lag_max = 10))
lines(1:10, acf_trawl(sim, sim$h, method = "acf", lag_max = 10), col = "blue")

# empirical and estimated autocorrelation for a differenced process
sim <- sim_trawl(list("levy_seed" = "Skellam", levy_par = c(0.13, 0.11), b = 0.3))
sim$h <- exp(seq(log(1e-2), log(60), length.out = 51))
sim$trawl <- "exp"
sim$include_b <- TRUE
sim$include_cum1 <- FALSE
sim$method <- "vs_C"
ftC <- fit_trawl(sim)
h <- 0.5
plot(1:10, acf_trawl(sim, h, dff = 1, lag_max = 10))
lines(1:10, acf_trawl(sim, h, method = "vs_C", dff = 1, lag_max = 10), col = "blue")

}
\references{
Barndorff‐Nielsen, O. E., Lunde, A., Shephard, N., & Veraart, A. E. (2014). 
Integer‐valued Trawl Processes: A Class of Stationary Infinitely Divisible Processes. 
Scandinavian Journal of Statistics, 41(3), 693-724.

Shephard, N., & Yang, J. J. (2017). 
Continuous time analysis of fleeting discrete price moves. 
Journal of the American Statistical Association, 112(519), 1090-1106.

Veraart, A. E. (2018). 
Modelling, simulation and inference for multivariate time series of counts. 
arXiv preprint arXiv:1608.03154.
}
\seealso{
\code{\link{fit_trawl}}, \code{\link{sim_trawl}}
}
\author{
Dries Cornilly
}
\concept{
trawl
}
