% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{ccf_sample}
\alias{ccf_sample}
\title{Sample ACF of irregularly observed time series}
\usage{
ccf_sample(object, h, dff = 0, lag_min = -25, lag_max = 25)
}
\arguments{
\item{\dots}{any other passthru pareters}
}
\description{
sample autocorrelation of a trawl process
}
\details{
TODO

CITE TODO.
}
\examples{

TODO

# simulations estimation
TODO

}
\references{
TODO
}
\seealso{
\code{\link{fit_trawl}}
}
\author{
Dries Cornilly
}
\concept{
trawl
}
