% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{ccf_trawl}
\alias{ccf_trawl}
\title{ACF of Trawl processes}
\usage{
ccf_trawl(object, h, dff = 0, lag_max = 25)
}
\arguments{
\item{\dots}{any other passthru pareters}
}
\description{
theoretical autocorrelation of a trawl process
}
\details{
TODO

CITE TODO.
}
\examples{

TODO

# simulations estimation
TODO

}
\references{
TODO
}
\seealso{
\code{\link{fit_trawl}}
}
\author{
Dries Cornilly
}
\concept{
trawl
}
