% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trawl.R
\name{fit_trawl}
\alias{fit_trawl}
\title{Estimation of Trawl processes}
\usage{
fit_trawl(object, ...)
}
\arguments{
\item{object}{bla}

\item{\dots}{any other passthru pareters}
}
\description{
estimates a path of a Trawl process
}
\details{
TODO

CITE TODO.
}
\examples{

#TODO

# simulations estimation
#TODO

}
\references{
TODO
}
\seealso{
\code{\link{fit_trawl}}
}
\author{
Dries Cornilly
}
\concept{
trawl
}
