% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trawl.R
\name{sim_trawl}
\alias{sim_trawl}
\title{Simulation of trawl processes}
\usage{
sim_trawl(object, univariate = TRUE, ...)
}
\arguments{
\item{object}{object containing all the specifications for the process, see details}

\item{univariate}{boolean indicating univariate or multivariate simulation}

\item{\dots}{any other passthrough parameters}
}
\value{
x_grid: vector of time points where the process value is given

p_grid: vector with process values for the time points in x_grid

the other return arguments echo the process specifications
}
\description{
simulation of univariate or multivariate trawl processes
}
\details{
the passed object contains all the process specifications in list format. The following
arguments should ideally be present in the object: levy seed (levy_seed), corresponding
levy parameters (levy_par), trawl (trawl), corresponding trawl parameters (trawl_par),
initial observation time (T0), end of observation period (TT), the frequency at which
the process is observed (observed_freq), and the b parameter governing the distinction
between levy and trawl process changes. In case of multivariate simulation, also the
design matrix showing the underlying factors should be provided (design_matrix).

The observation frequency is for example 1 second, 
which means that a change in the process value is only observed the next round second. 
For the b parameter, we refer to Shephard and Yang (2017). Some default parameters are set,
but it is best not to rely on these. See examples.
}
\examples{

# default settings - Poisson process with exponential trawl
sim <- sim_trawl(list())

# Skellam process with gig trawl
sim <- sim_trawl(list("levy_seed" = "Skellam", "levy_par" = c(0.131, 0.130),
                      "trawl" = "gig", "trawl_par" = c(0.01, 0.45, -0.6),
                      "T0" = 72.03, "TT" = 75600, "observed_freq" = 1e-6, 
                      "b" = 0.3))

# multivariate Poisson / Skellam with gamma and exponential trawls
trawl <- list("gamma", "exp")
trawl_par <- list(c(0.9, 1.8), c(0.5))
levy_par <- matrix(c(0.13, 0.13, 0.23, 0.11, 0.05), ncol = 1)
design_matrix <- matrix(c(1, 0, 0, 1, -1, 1, 1, -1, 1, 0), nrow = 2)
sim <- sim_trawl(list("levy_seed" = "Poisson", "levy_par" = levy_par,
                      "trawl" = trawl, "trawl_par" = trawl_par,
                      "design_matrix" = design_matrix, "b" = c(0, 0),
                      "T0" = 0, "TT" = 75600, "observed_freq" = 1e-6), 
                 univariate = FALSE)

# multivariate negative binomial
trawl <- list("gamma", "exp")
trawl_par <- list(c(0.9, 1.8), c(0.5))
levy_par <- matrix(c(1.5, 0.4, 0.3,
                     0.7, 0.4, NA,
                     0.6, 0.3, NA), ncol = 3, byrow = TRUE)
design_matrix <- matrix(c(1, 1, 1, 0, 0, 1), nrow = 2)
sim <- sim_trawl(list("levy_seed" = "negBin", "levy_par" = levy_par,
                      "trawl" = trawl, "trawl_par" = trawl_par,
                      "design_matrix" = design_matrix, "b" = c(0, 0),
                      "T0" = 15.3, "TT" = 35600, "observed_freq" = 1e-3), 
                 univariate = FALSE)

}
\references{
Barndorff‐Nielsen, O. E., Lunde, A., Shephard, N., & Veraart, A. E. (2014). 
Integer‐valued Trawl Processes: A Class of Stationary Infinitely Divisible Processes. 
Scandinavian Journal of Statistics, 41(3), 693-724.

Shephard, N., & Yang, J. J. (2017). 
Continuous time analysis of fleeting discrete price moves. 
Journal of the American Statistical Association, 112(519), 1090-1106.

Veraart, A. E. (2018). 
Modelling, simulation and inference for multivariate time series of counts. 
arXiv preprint arXiv:1608.03154.
}
\seealso{
\code{\link{fit_trawl}}
}
\author{
Dries Cornilly
}
\concept{
trawl
}
