% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_trawl.R
\name{vs_trawl}
\alias{vs_trawl}
\title{Estimation of Trawl processes}
\usage{
vs_trawl(object, h, method = "sample", ...)
}
\arguments{
\item{object}{bla}

\item{h}{bla}

\item{method}{bla}

\item{\dots}{any other passthru pareters}
}
\description{
display variance signature plot
}
\details{
TODO

CITE TODO.
}
\examples{

#TODO

# simulations estimation
#TODO

}
\references{
TODO
}
\seealso{
\code{\link{fit_trawl}}
}
\author{
Dries Cornilly
}
\concept{
trawl
}
